Created user-secrets for development use.
Created dotnet user-secrets for development use. No sensitive
information will need to be commited to github. Started to build out
unit tests to untilize a mongodbSettings.json that will not be sourced
as well. The plan is to have the connection string as an Azure Devops
variable and pass in those arugments for CI/CD.