Added Proper DI, proper switching between json configs, begain normalizing user secrets and env vars for local testing and CI/CD
There were numerous additions to this commit. Mainly for configuration of getting WorkerService resolved through proper DI. Also started to normalize the user-secrets and env vars. It was fragemented until this point. There was an issue getting Azure envs set and read in the Unit Tests. This was due to not using all caps for env vars (duh, idiot). This is the first step in getting CI/CD fully funtional. Next commit all configuration, user-secrets, env vars should be 100% set and normalized for full rapid developement